@@ -1,42 +1,40 @@
 {
-    "name": "bittorrent-utils",
-    "version": "2.5.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-        "dev": "cross-env NODE_ENV=dev nodemon index.js",
-        "dev-manual-transfer": "cross-env NODE_ENV=dev nodemon ./src/manualTransfer.js",
-        "start": "npm install --omit=dev && node -e \"console.clear()\" && cross-env NODE_ENV=user node index.js",
-        "start-manual-transfer": "npm install --omit=dev && node -e \"console.clear()\" && cross-env NODE_ENV=user node ./src/manualTransfer.js",
-        "lint": "prettier --check . && eslint **/*.js",
-        "format": "prettier --write . && eslint --fix **/*.js"
-    },
-    "author": "Yaroslav Sorochan <yaroslav.sorochan.i@gmail.com>",
-    "license": "MIT",
-    "devDependencies": {
-        "eslint": "^8.18.0",
-        "eslint-config-prettier": "^8.5.0",
-        "eslint-config-standard": "^17.0.0",
-        "eslint-plugin-import": "^2.26.0",
-        "eslint-plugin-n": "^15.2.3",
-        "eslint-plugin-promise": "^6.0.0",
-        "nodemon": "^2.0.12",
-        "prettier": "2.7.1"
-    },
-    "dependencies": {
-        "@grpc/grpc-js": "^1.3.7",
-        "@grpc/proto-loader": "^0.6.4",
-        "@influxdata/influxdb-client": "^1.27.0",
-        "b58": "^4.0.3",
-        "colors": "^1.4.0",
-        "config": "^3.3.6",
-        "cross-env": "^7.0.3",
-        "keccak": "^3.0.2",
-        "loglevel": "^1.7.1",
-        "node-fetch": "^2.6.1",
-        "parent-module": "^2.0.0",
-        "protobufjs": "^6.11.2",
-        "secp256k1": "^4.0.2",
-        "semver": "^7.3.5"
-    }
+  "name": "bittorrent-utils",
+  "version": "3.0.0",
+  "description": "",
+  "main": "index.js",
+  "type": "module",
+  "scripts": {
+    "start": "npm install --omit=dev && node -e \"console.clear()\" && cross-env NODE_ENV=user node index.js",
+    "dev": "cross-env NODE_ENV=dev node --watch index.js",
+    "start-manual-transfer": "npm install --omit=dev && node -e \"console.clear()\" && cross-env NODE_ENV=user node ./src/modules/manualTransfer.js",
+    "dev-manual-transfer": "cross-env NODE_ENV=dev node --watch ./src/modules/manualTransfer.js",
+    "lint": "eslint --fix --ext .js . && prettier --write ."
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "eslint": "^8.26.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.4.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "prettier": "^2.7.1"
+  },
+  "dependencies": {
+    "@grpc/grpc-js": "^1.7.0",
+    "@grpc/proto-loader": "^0.7.2",
+    "@influxdata/influxdb-client": "^1.29.0",
+    "b58": "^4.0.3",
+    "colors": "^1.4.0",
+    "cross-env": "^7.0.3",
+    "find-process": "^1.4.7",
+    "keccak": "^3.0.2",
+    "loglevel": "^1.8.0",
+    "protobufjs": "^7.1.1",
+    "secp256k1": "^4.0.3",
+    "semver": "^7.3.7"
+  }
 }
