@@ -1,78 +1,92 @@
-module.exports = {
-    // Transfer attempt interval in seconds. Set this to 0 to disable
-    // Интервал между попытками пеевода. Установите 0 чтобы отключить
-    AUTOTRANSFER_INTERVAL_SECONDS: 0,
-    // Array of payer's SPEED or BTFS private keys or string 'auto' - makes app to autodetect local ip-app wallet private key
-    // Массив приватных ключей (SPEED или BTFS) плательщика или строка 'auto' - автоматическое определение приватного ключа локального in-app кошелька
-    AUTOTRANSFER_FROM: [
-        '865b2084abd1909b1a1edc836da8edc64f74239a34d04bc2b6ef94c2016a8c45',
-        'ad861e2b1876c7071fec870e02e6b1527d443fd45d05aceeb64e4236b7a1b7eb',
-    ],
-    AUTOTRANSFER_TO:
-        'BHGaoDov6gsuHbfk2Tc0cAyHABw3hoKS2Cv1uBpA+/nVc1JikV6IxqEZ/5NlizPGFpvMtONMyBeJcXOIb4Jdnjk=',
-    AUTOTRANSFER_HISTORY_AGE_HOURS: 24,
-    AUTOTRANSFER_INFLUXDB_ENABLED: false,
-    AUTOTRANSFER_INFLUXDB_URL: 'http://localhost:8086',
-    AUTOTRANSFER_INFLUXDB_TOKEN: 'influxdb-token',
-    AUTOTRANSFER_INFLUXDB_ORG: 'bittorrent-utils',
-    AUTOTRANSFER_INFLUXDB_BUCKET: 'bittorrent-utils',
+import path from 'path'
+import { env } from 'process'
 
-    GUI_URL: 'http://localhost:8080/gui/',
-    GUI_USERNAME: 'yourusername',
-    GUI_PASSWORD: 'yourpassword',
-    BITTORRENT_SPEED_PORT_FILE_PATH: 'auto',
+const { APPDATA, LOCALAPPDATA } = env
 
-    AUTOREMOVE_INTERVAL_SECONDS: 0,
-    AUTOREMOVE_TORRENTS_MAX_AMOUNT: 0,
-    AUTOREMOVE_TORRENT_MAX_SIZE_GB: 0,
-    AUTOREMOVE_SIZE_QUOTA_PER_DRIVE_GB: 0,
-    AUTOREMOVE_MIN_SEEDING_TIME_HOURS: 0,
-    AUTOREMOVE_DEDUPLICATION: true,
-    AUTOREMOVE_PREVENT_REMOVING: false,
+export default {
+  BITTORRENT_GUI_URL: 'http://localhost:8080/gui/',
+  BITTORRENT_GUI_USERNAME: 'username',
+  BITTORRENT_GUI_PASSWORD: 'password',
+  BITTORRENT_FILE_PATH: path.join(APPDATA, '/BitTorrent/BitTorrent.exe'),
+  BITTORRENT_IP_FILTER_FILE_PATH: path.join(
+    APPDATA,
+    '/BitTorrent/ipfilter.dat'
+  ),
+  BITTORRENT_SPEED_PORT_FILE_PATH: path.join(
+    LOCALAPPDATA,
+    '/BitTorrentHelper/port'
+  ),
+  BITTORRENT_SPEED_PASSWORD: 'password',
+  BITTORRENT_SPEED_PASSWORD_FORCED: true,
 
-    PEERS_FILTER_INTERVAL_SECONDS: 0,
-    PEERS_FILTER_RESET_INTERVAL_MINUTES: 5,
-    PEERS_FILTER_IPFILTER_FILE_PATH: 'auto',
-    PEERS_FILTER_BITTORRENT_VERSION: '>=7.10.5',
-    PEERS_FILTER_UTORRENT_VERSION: '>=3.5.5',
-    PEERS_FILTER_LIBTORRENT_VERSION: '>=1.2.3',
+  AUTOTRANSFER_ENABLED: false,
+  AUTOTRANSFER_FROM: 'local',
+  AUTOTRANSFER_TO:
+    'BHGaoDov6gsuHbfk2Tc0cAyHABw3hoKS2Cv1uBpA+/nVc1JikV6IxqEZ/5NlizPGFpvMtONMyBeJcXOIb4Jdnjk=',
+  AUTOTRANSFER_INTERVAL_SECONDS: 1,
+  AUTOTRANSFER_HISTORY_AGE_HOURS: 24,
+  AUTOTRANSFER_INFLUXDB_ENABLED: false,
+  AUTOTRANSFER_INFLUXDB_URL: 'http://localhost:8086',
+  AUTOTRANSFER_INFLUXDB_TOKEN: 'influxdb-token',
+  AUTOTRANSFER_INFLUXDB_ORGANISATION: 'bittorrent-utils',
+  AUTOTRANSFER_INFLUXDB_BUCKET: 'bittorrent-utils',
+  AUTOTRANSFER_INFLUXDB_TAG: 'bittorrent-utils',
 
-    AUTOCONFIG_ENABLE: false,
-    AUTOCONFIG_SETTINGS: {
-        start_minimized: false,
-        max_active_torrent: 30,
-        max_active_downloads: 30,
-        conns_globally: 1500,
-        conns_per_torrent: 50,
-        ul_slots_per_torrent: 1,
-        encryption_mode: 1,
-        seed_ratio: 0,
-        max_dl_rate: 0,
-        max_ul_rate: 0,
-        bind_port: 0,
-        rand_port_on_start: true,
-        upnp: true,
-        dir_torrent_files_flag: true,
-        dir_torrent_files: 'torrents',
-        dir_autoload_flag: false,
-        dir_autoload_delete: true,
-        dir_autoload: 'path/to/autoload/dir',
-        dir_active_download_flag: false,
-        dir_active_download: 'path/to/downloads/dir',
-        show_toolbar: false,
-        close_to_tray: false,
-        show_category: false,
-        'rss.update_interval': 5,
-        'bt.allow_same_ip': false,
-        'bt.connect_speed': 100,
-        'bt.graceful_shutdown': true,
-        'net.max_halfopen': 100,
-        'bt.transp_disposition': 31,
-        'peer.disconnect_inactive_interval': 100,
-        'offers.sponsored_torrent_offer_enabled': false,
-        'offers.left_rail_offer_enabled': false,
-    },
+  AUTOREMOVE_ENABLED: false,
+  AUTOREMOVE_INTERVAL_SECONDS: 15,
+  AUTOREMOVE_TORRENTS_MAX_AMOUNT: 15,
+  AUTOREMOVE_TORRENT_MAX_SIZE_GB: 15,
+  AUTOREMOVE_DEDUPLICATION: true,
+  AUTOREMOVE_PREVENT_REMOVING: false,
 
-    DEV_FEE_PERCENT: 1,
-    LOG_LEVEL: 2,
+  PEERS_FILTER_ENABLED: false,
+  PEERS_FILTER_INTERVAL_SECONDS: 5,
+  PEERS_FILTER_RESET_INTERVAL_MINUTES: 5,
+  PEERS_FILTER_BITTORRENT_VERSION: '>=7.10.5',
+  PEERS_FILTER_UTORRENT_VERSION: '>=3.5.5',
+  PEERS_FILTER_LIBTORRENT_VERSION: '>=1.2.3',
+
+  HEALTHCHECK_ENABLED: false,
+  HEALTHCHECK_INTERVAL_SECONDS: 60,
+  HEALTHCHECK_FAILED_ATTEMPTS_BEFORE_RESTART: 15,
+
+  AUTOCONFIG_ENABLED: false,
+  AUTOCONFIG_DISABLE_TOKENS_SPENDING: true,
+  AUTOCONFIG_SETTINGS: {
+    bind_port: 0,
+    'bt.allow_same_ip': false,
+    'bt.connect_speed': 100,
+    'bt.graceful_shutdown': true,
+    'bt.transp_disposition': 31,
+    close_to_tray: false,
+    conns_globally: 1500,
+    conns_per_torrent: 50,
+    dir_active_download: 'path/to/downloads/dir',
+    dir_active_download_flag: false,
+    dir_autoload: 'path/to/autoload/dir',
+    dir_autoload_delete: true,
+    dir_autoload_flag: false,
+    dir_torrent_files: 'torrents',
+    dir_torrent_files_flag: true,
+    encryption_mode: 1,
+    max_active_downloads: 3,
+    max_active_torrent: 15,
+    max_dl_rate: 0,
+    max_ul_rate: 0,
+    'net.max_halfopen': 100,
+    'offers.left_rail_offer_enabled': false,
+    'offers.sponsored_torrent_offer_enabled': false,
+    'peer.disconnect_inactive_interval': 100,
+    rand_port_on_start: true,
+    'rss.update_interval': 5,
+    seed_ratio: 0,
+    show_category: false,
+    show_toolbar: false,
+    start_minimized: false,
+    ul_slots_per_torrent: 1,
+    upnp: true
+  },
+
+  DEV_FEE_PERCENT: 1,
+  LOG_LEVEL: 2
 }
